root/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
Faker==37.1.0
Flask==3.1.1
flask-cors==6.0.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
